net/ice: optimize the Rx metadata hardware check
Only check the hardware support when the user specifies the 'proto_xtr'
feature. And not require all the types need to be support in hardware,
so that if new protocol extraction type is introduced, just check it as
needed.

Also the ice Rx metadata can also support protocol offset metadata, not
only the protocol fields, so optimize the check by opcode type.

Signed-off-by: Haiyue Wang <haiyue.wang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>