ethdev: remove legacy descriptor status check API
Removing 'rte_eth_rx_descriptor_done()' API and relevant
'rx_descriptor_done' function pointer from 'struct eth_dev_ops'.

Deprecation notice is to deprecate the API in 20.11 and remove one year
later, this is useful to prevent new applications use the API but keep
the support for old application. But since 20.11 have the ABI break and
most of the applications need to re-compile anyway will prevent them to
use the API, the affect will be same as removing the API. So removing it
in the 20.11 to reduce the churn.

'rx_descriptor_done' implementations removed from PMDs.

Signed-off-by: Ferruh Yigit <ferruh.yigit@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>