config: remove explicit undefinition of unset values
Rather than explicitly clearing any setting of undefined values in our
rte_config.h file, it's better to instead just add a comment that the value
is not set. Using a comment allows the user to set the value using CFLAGS
or similar mechanism without the config file clearing the value again.

The text used "<VALUE> is not set" is modelled after the kernel approach of
doing the same thing.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>