app/testpmd: fix name of bitrate library in meson build
The bitrate library in DPDK is actually in a "bitratestats" directory, so
that is used by meson for the macro and library name. Therefore, we need
to update references to RTE_LIBRTE_BITRATE to RTE_LIBRTE_BITRATESTATS in
testpmd to have it found. Rather than supporting both defines, since make
is being removed, we can just replace all instances of the former define
with the latter.

To ensure testpmd links ok when this is done, we also need to add
bitratestats to the list of library dependencies.

Fixes: 5b9656b157d3 ("lib: build with meson")
Cc: stable@dpdk.org

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>