net/iavf: fix incorrect link status when speed is undefined
When PF is put in no-carrier state, iavf VFs will switch to
"in carrier" state due to a link up + a link speed set to 0
(default value if no speed detected).

To be consistent with linux drivers on which PF and VFs are in
same carrier state, updates a link status of VF only if link is up
and speed is different from undefined.

Fixes: 48de41ca11f0 ("net/avf: enable link status update")

Signed-off-by: SteveX Yang <stevex.yang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>