gro: add UDP GRO support
UDP GRO can help improve VM-to-VM UDP performance when
VM is enabled UFO or GSO, GRO must be supported if GSO
or UFO is enabled, otherwise, performance gain will be
hurt.

With this enabled in DPDK, OVS DPDK can leverage it
to improve VM-to-VM UDP performance, this will make
sure IP fragments will be reassembled once it is
received from physical NIC. It is very helpful in OVS
DPDK VLAN TSO case.

Signed-off-by: Yi Yang <yangyi01@inspur.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>