net/dpaa2: fix build error about timesync functions
When the build option has '-DRTE_LIBRTE_IEEE1588=1',
the announce of timesync functions will be build.
However the dpdk_conf doesn't hav RTE_LIBRTE_IEEE1588
so that the file dpaa2_ptp.c will not be build.
It cause the build error.
This patch fixes it by adding set for dpdk_conf.

Fixes: 184c39d16568 ("net/dpaa2: add DPRTC sub-module")
Cc: stable@dpdk.org

Signed-off-by: Chenxu Di <chenxux.di@intel.com>
Acked-by: Hemant Agrawal <hemant.agrawal@nxp.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>